remove more warnings on windows

+++ Code/GraphMol/FileParsers/MolFileStereochem.cpp
@@ -103,7 +103,7 @@ Atom::ChiralType FindAtomStereochemistry(const RWMol &mol, const Bond *bond,
     }
     ++beg;
   }
-  int nNbrs = neighborBondIndices.size();
+  size_t nNbrs = neighborBondIndices.size();
 
   //----------------------------------------------------------
   //
